LayoutTests/fast/dom/htmlcollection-zombies-expected.txt

This test tests the behavior of collections after their base node has been destroyed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing selectElement's options collection.
PASS collection.length is 2
PASS collection[0].getAttribute('name') is 'foo'
PASS collection[1].getAttribute('name') is 'bar'
Destroying selectElement
PASS collection.length is 0
PASS collection[0] is undefined
Testing formElement's elements collection.
PASS collection.length is 2
PASS collection[0].getAttribute('name') is 'foo'
PASS collection[1].getAttribute('name') is 'bar'
Destroying formElement
PASS collection.length is 0
PASS collection[0] is undefined
Testing tableElement's rows collection.
PASS collection.length is 2
PASS collection[0].getAttribute('name') is 'foo'
PASS collection[1].getAttribute('name') is 'bar'
Destroying tableElement
PASS collection.length is 0
PASS collection[0] is undefined
PASS successfullyParsed is true

TEST COMPLETE
